requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
bleach==6.2.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-tailwind==1.0.3
cryptography==44.0.2
Django==5.0.6
django-allauth==0.65.7
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-crispy-forms==2.4
django-environ==0.12.0
django-htmx==1.23.0
django-js-asset==3.1.2
django-summernote==0.8.20.0
django-tailwind==3.8.0
django_celery_results==2.6.0
idna==3.10
joblib==1.4.2
kombu==5.5.3
lxml==5.4.0
nltk==3.9.1
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
redis==6.0.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
